+++ src/ui.rs
@@ -5,7 +5,7 @@ use sdl2::video::Window;
 
 const CELL_SIZE: i32 = 30;
 const CELL_PADDING: i32 = 2;
-const BUFFER_SIZE: i32 = 60;
+pub const BUFFER_SIZE: i32 = 60;
 const MINIMAL_SCREEN_SIZE: u32 = 2 * BUFFER_SIZE as u32 + CELL_SIZE as u32;
 
 /// Struct to convert between grid coordinates and screen coordinates
@@ -134,131 +134,5 @@ pub fn render_grid(canvas: &mut Canvas<Window>) -> Result<(i32, i32), String> {
     }
 
     Ok((rows, cols))
-}
-
-/// Given three points, render the outline of a triangle abc
-pub fn draw_triangle(canvas: &mut Canvas<Window>, a: Vector2, b: Vector2, c: Vector2) {
-    canvas.draw_line(a, b).unwrap();
-    canvas.draw_line(b, c).unwrap();
-    canvas.draw_line(c, a).unwrap();
-}
-
-/// Given two points that form a line, and a y-coordinate, interpolate the x-coordinate 
-fn interpolate(a: Vector2, b: Vector2, y: i32) -> i32 {
-    let m = (b.y - a.y) as f32 / (b.x - a.x) as f32;
-    ((y - a.y) as f32 / m) as i32 + a.x
-} 
-
-/// Given three points, fill the triangle abc using a scanline algorithm
-pub fn fill_triangle(canvas: &mut Canvas<Window>, a: Vector2, b: Vector2, c: Vector2) {
-    let mut points: Vec<Vector2> = vec![a, b, c];
-    points.sort_by(|first, second| first.y.cmp(&second.y));
-
-    // iterate through each y-level
-    for y in points[0].y..=points[2].y {
-        // compute active edges
-        let left_edge: (Vector2, Vector2);
-        let right_edge: (Vector2, Vector2);
-
-        /*
-        configuration where the intermediate point is to the right of the lowest point:
-         -
-         |\
-         | \
-         |  -
-         | /
-         |/
-         -
-         */
-        if points[0].x <= points[1].x && y <= points[1].y {
-            left_edge = (points[0], points[2]);
-            right_edge = (points[0], points[1]);
-        } else if points[0].x <= points[1].x {
-            left_edge = (points[0], points[2]);
-            right_edge = (points[1], points[2]);
-        }
-
-        /*
-        configuration where the intermediate point is to the left of the lowest point:
-            -
-           /|
-          / |
-         -  |
-          \ |
-           \|
-            -
-         */
-        else if y <= points[1].y {
-            left_edge = (points[0], points[1]);
-            right_edge = (points[0], points[2]);
-        } else {
-            left_edge = (points[1], points[2]);
-            right_edge = (points[0], points[2]);
-        }
-
-        // get start and end points of line to draw
-        let start: Vector2 = Vector2::new(interpolate(left_edge.0, left_edge.1, y), y);
-        let end: Vector2 = Vector2::new(interpolate(right_edge.0, right_edge.1, y), y);
-
-        // draw the line
-        canvas.draw_line(start, end).unwrap();
-    }
-}
-
-/// Render a play button in the bottom buffer
-pub fn render_play(canvas: &mut Canvas<Window>) {
-    // get screen size and set draw color
-    let screen_size: (u32, u32) = canvas.output_size().unwrap();
-    let screen_width = screen_size.0 as i32;
-    let screen_height = screen_size.1 as i32;
-    canvas.set_draw_color(Color::BLACK);
-
-    // dimensions for the play button triangle, using equilateral triangle lengths
-    let height: i32 = BUFFER_SIZE / 2;
-    let width: i32 = ((height as f32 / 2.0) * (3 as f32).sqrt()) as i32;
-
-    // padding on top and on the bottom for the play button triangle
-    let p_top: i32 = (BUFFER_SIZE - height) / 2;
-    let p_bottom: i32 = BUFFER_SIZE - p_top - height;
-
-    // points for the play button triangle
-    let a = Vector2::new((screen_width - width) / 2, screen_height - BUFFER_SIZE + p_top);
-    let b = Vector2::new((screen_width - width) / 2, screen_height - p_bottom);
-    let c = Vector2::new((screen_width + width) / 2, screen_height - p_bottom - height / 2);
-
-    // render the triangle
-    fill_triangle(canvas, a, b, c);
-}
-
-/// Render a pause button in the bottom buffer
-pub fn render_pause(canvas: &mut Canvas<Window>) { 
-    // get screen size and set draw color
-    let screen_size: (u32, u32) = canvas.output_size().unwrap();
-    let screen_width = screen_size.0 as i32;
-    let screen_height = screen_size.1 as i32;
-    canvas.set_draw_color(Color::BLACK);
-
-    // dimensions for the pause button, including distance between the two rectangles
-    let height: i32 = BUFFER_SIZE / 2;
-    let width: i32 = height / 4;
-    let distance: i32 = height / 3;
-
-    // padding on top and on the bottom for the pause button
-    let p_top: i32 = (BUFFER_SIZE - height) / 2;
-    let p_bottom: i32 = BUFFER_SIZE - p_top - height;
-
-    // create the rectangles
-    let left_rect = Rect::new((screen_width - distance) / 2 - width,
-        screen_height - p_bottom - height,
-        width as u32,
-        height as u32);
-    let right_rect = Rect::new((screen_width + distance) / 2,
-        screen_height - p_bottom - height,
-        width as u32,
-        height as u32);
-
-    // render the rectangles
-    canvas.fill_rect(left_rect).unwrap();
-    canvas.fill_rect(right_rect).unwrap();
 }
